Fixed bad calculation bug that could mess up timing once every 1.2hr of daemon uptime. Most likely the reason for why learn button stopped working after a while (and restarted working after another while...), and for a number of other occasional hickups we've seen in dimming, button state machine etc.
Background: struct timespec members are 32bit, and need cast to 64bit before multiplying with 1e6.
Without this, MainLoop::now() had a wraparound every 4294 seconds.

+++ src/p44utils/mainloop.cpp
@@ -38,7 +38,7 @@ MLMicroSeconds MainLoop::now()
   struct timespec tsp;
   clock_gettime(CLOCK_MONOTONIC, &tsp);
   // return milliseconds
-  return tsp.tv_sec*1000000 + tsp.tv_nsec/1000; // uS
+  return ((uint64_t)(tsp.tv_sec))*1000000ll + tsp.tv_nsec/1000; // uS
 #endif
 }
